feat(Redis): add `db` configuration option
Support for `redis-namespace` was [removed in Sidekiq 7][s7].

In response, have `froman` support `db` as an alternative to
`namespace` for checking if there is `work_to_do`

As far as I could tell, there's no way to access multiple dbs with the
same connection. So the major change here is to `runner.rs` where we
create a separate `redis_conn` for each `Worker`.

[s7]: https://www.mikeperham.com/2022/10/27/introducing-sidekiq-7.0/